# OnlineShopBackend-main/src/main/resources/static/swagger.yml
openapi: 3.0.3
info:
  title: Online Shop
  description: |-
    Aplicación que se encarga de habilitar endpoints para el consumo REST de la base de datos en la tienda online.
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: http://localhost:8080/api/v1
    description: Local server
...
paths:
...
  /auth/register:
    post:
      tags:
        - auth
      summary: Register a user
      description: Register a user and returns an access token and refresh token.
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRequest'
        required: true
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Invalid input
        '422':
          description: Validation exception
  /auth/login:
    post:
      tags:
        - auth
      summary: Log in a user
      description: Authenticates a user and returns an access token and refresh token.
      operationId: loginUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResponse'
        '400':
          description: Invalid input
        '401':
          description: Authentication failed
  /auth/verify/{token}:
    post:
      tags:
        - auth
      summary: Verificar la cuenta de un usuario usando el token de verificación
      description: Verifica la cuenta del usuario utilizando el token de verificación proporcionado.
      operationId: verifyUser
      parameters:
        - name: token
          in: path
          description: El token de verificación para verificar la cuenta
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Cuenta verificada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResponse'
        '400':
          description: Token de verificación inválido
          content:
            application/json:
              schema:
                type: string
                example: "Token de verificación inválido."
        '404':
          description: Token de verificación no encontrado
          content:
            application/json:
              schema:
                type: string
                example: "Token de verificación no encontrado."
  /auth/authenticated-user:
    get:
      tags:
        - auth
      summary: Obtener la información del usuario autenticado
      description: Recupera la información del usuario autenticado.
      operationId: getAuthenticatedUser
      responses:
        '200':
          description: Usuario autenticado encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                type: string
                example: "No se encontró ningún usuario autenticado."
  /auth/password-reset/request:
    post:
      tags:
        - auth
      summary: Envia el link de resetear la contraseña
      description: Envia el codigo de resetear la contraseña al email del usuario
      operationId: passwordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  description: The email address of the user.
                  example: user@example.com
      responses:
        '200':
          description: Password reset link sent successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailResponse'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                type: string
                example: "No se encontró ningún usuario autenticado."
  /auth/password-reset/confirm:
    patch:
      tags:
        - auth
      summary: Resetea la contrasela utilizando el token
      description: Si el token es valido, resetea la contraseña
      operationId: passwordResetConfirm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                verification-token:
                  type: string
                  description: The password reset token sent to the user's email.
                  example: dummytoken123
                newPassword:
                  type: string
                  format: password
                  description: The new password for the user.
                  example: P@ssw0rd123
      responses:
        '200':
          description: Password reset successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '400':
          description: Invalid or expired token.
          content:
            application/json:
              schema:
                type: string
                example: "No se encontró ningún usuario autenticado."
...
components:
  schemas:
...
    EmailResponse:
      type: object
      properties:
        message:
          type: string
...
    CustomerRequest:
      type: object
      properties:
        name:
          type: string
          example: John
        password:
          type: string
          example: '12345'
        surname:
          type: string
          example: James
        surname2:
          type: string
          example: James
        address:
          type: string
          example: Calle nombre n3 1b
        city:
          type: string
          example: Chiclana de la Frontera
        province:
          type: string
          example: Cadiz
        region:
          type: string
          example: España
        postalCode:
          type: string
          example: "11130"
        email:
          type: string
          example: john@email.com
        phone:
          type: string
          example: '12345'
        status:
          type: boolean
          description: Estado del usuario
          example: true
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          example: john@email.com
        password:
          type: string
          example: '12345'
    CustomerResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        username:
          type: string
          example: theUser
        password:
          type: string
          example: '12345'
        name:
          type: string
          example: John
        surname:
          type: string
          example: James
        surname2:
          type: string
          example: James
        address:
          type: string
          example: Calle nombre n3 1b
        city:
          type: string
          example: Chiclana de la Frontera
        province:
          type: string
          example: Cadiz
        region:
          type: string
          example: España
        postalCode:
          type: string
          example: "11130"
        email:
          type: string
          example: john@email.com
        phone:
          type: string
          example: '12345'
        status:
          type: boolean
          description: Estado del usuario
          example: true
        rol:
          type: string
          format: enum
            - USER
            - ADMIN
          description: Rol de usuario
          example: USER
    AuthenticationResponse:
      type: object
      properties:
        accessToken:
          type: string
          example: eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQSflwTmsuHo_iij07PpRZsI7UlFz06OZtToOZfpC9q6U
        refreshToken:
          type: string
          example: dGhpcy1yZWZyZXNoLXRva2Vu
        message:
          type: string
          example: Authentication successful
    RegisterResponse:
      type: object
      properties:
        accessToken:
          type: string
          example: null
        refreshToken:
          type: string
          example: null
        message:
          type: string
          example: Usuario registrado correctamente. Por favor revisa tu correo electrónico para verificar la cuenta.
...